LOGOUT fertig; DECKBUILDER begonnen

+++ CardGameLap/CardGame/CardGame.Web/Models/Card.cs
@@ -13,13 +13,13 @@ namespace CardGame.Web.Models
         public string Name { get; set; }
 
         [Range(0,int.MaxValue)]
-        public int Mana { get; set; }
+        public byte Mana { get; set; }
 
         [Range(0, int.MaxValue)]
-        public int Attack { get; set; }
+        public short Attack { get; set; }
 
         [Range(0, int.MaxValue)]
-        public int Life { get; set; }
+        public short Life { get; set; }
         
         public string Type { get; set; }
